# docs/swagger.yaml
basePath: /
definitions:
...
  registration.CreateCheckInAgentRequest:
    properties:
      name:
        type: string
    type: object
  registration.RegistrationStatus:
    enum:
    - registered
    - checkedin
    - cancelled
    type: string
    x-enum-varnames:
    - RegistrationStatusRegistered
    - RegistrationStatusCheckedIn
    - RegistrationStatusCancelled
  registration.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      otp_required:
        type: boolean
      status:
        example: success
        type: string
    type: object
  registration.UpdateRegistrationStatusRequest:
    properties:
      formId:
        type: string
      regId:
        type: integer
      regRef:
        type: string
      regStatus:
        $ref: '#/definitions/registration.RegistrationStatus'
      regType:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@gbaski.app
    name: API Support
  description: API documentation for Gbaski Host service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gbaski Host API
  version: "1.0"
paths:
...
  /registrations/check-in:
    get:
      consumes:
      - application/json
      description: Check in a registration using registration reference and agent
        token
      parameters:
      - description: Registration reference
        in: query
        name: regRef
        required: true
        type: string
      - description: Bearer token (check-in agent token)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/registration.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/registration.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/registration.Response'
      summary: Check in registration
      tags:
      - registrations
...
  /registrations/lookup:
    get:
      consumes:
      - application/json
      description: Lookup a registration by registration reference using agent token
      parameters:
      - description: Registration reference
        in: query
        name: regRef
        required: true
        type: string
      - description: Bearer token (check-in agent token)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/registration.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/registration.Response'
      summary: Lookup registration
      tags:
      - registrations
  /registrations/ticket/{formId}:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of ticket registrations for a specific
        form
      parameters:
      - description: Form ID (UUID)
        in: path
        name: formId
        required: true
        type: string
      - description: Search query
        in: query
        name: search
        type: string
      - description: Filter parameters
        in: query
        name: filter
        type: string
      - description: Range parameters
        in: query
        name: range
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/registration.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/registration.Response'
      security:
      - Bearer: []
      summary: Get ticket registrations
      tags:
      - registrations
  /registrations/ticket/{formId}/attendees:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of ticket attendees for a specific form
      parameters:
      - description: Form ID (UUID)
        in: path
        name: formId
        required: true
        type: string
      - description: Search query
        in: query
        name: search
        type: string
      - description: Filter parameters
        in: query
        name: filter
        type: string
      - description: Range parameters
        in: query
        name: range
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/registration.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/registration.Response'
      security:
      - Bearer: []
      summary: Get ticket attendees
      tags:
      - registrations
  /registrations/ticket/{formId}/check-in-agents:
    get:
      consumes:
      - application/json
      description: Retrieve all check-in agents for a specific form
      parameters:
      - description: Form ID (UUID)
        in: path
        name: formId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/registration.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/registration.Response'
      security:
      - Bearer: []
      summary: Get ticket check-in agents
      tags:
      - registrations
    post:
      consumes:
      - application/json
      description: Create a new check-in agent for a specific form
      parameters:
      - description: Form ID (UUID)
        in: path
        name: formId
        required: true
        type: string
      - description: Check-in agent creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/registration.CreateCheckInAgentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/registration.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/registration.Response'
      security:
      - Bearer: []
      summary: Create ticket check-in agent
      tags:
      - registrations
  /registrations/ticket/{formId}/check-in-agents/{token}:
    delete:
      consumes:
      - application/json
      description: Delete a check-in agent by token for a specific form
      parameters:
      - description: Form ID (UUID)
        in: path
        name: formId
        required: true
        type: string
      - description: Agent token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/registration.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/registration.Response'
      security:
      - Bearer: []
      summary: Delete ticket check-in agent
      tags:
      - registrations
  /registrations/ticket/{formId}/stats:
    get:
      consumes:
      - application/json
      description: Retrieve statistics for ticket registrations of a specific form
      parameters:
      - description: Form ID (UUID)
        in: path
        name: formId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/registration.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/registration.Response'
      security:
      - Bearer: []
      summary: Get ticket registration statistics
      tags:
      - registrations
  /registrations/ticket/{formId}/status:
    put:
      consumes:
      - application/json
      description: Update the status of a ticket registration (check-in)
      parameters:
      - description: Form ID (UUID)
        in: path
        name: formId
        required: true
        type: string
      - description: Registration status update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/registration.UpdateRegistrationStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/registration.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/registration.Response'
      security:
      - Bearer: []
      summary: Update ticket registration status
      tags:
      - registrations
  /registrations/ticket/check-in-agent/{state}:
    get:
      consumes:
      - application/json
      description: Retrieve a check-in agent by state token
      parameters:
      - description: Agent state token
        in: path
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/registration.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/registration.Response'
      summary: Get ticket check-in agent
      tags:
      - registrations
...
schemes:
- http
- https
swagger: "2.0"